+++ CHANGELOG.md
@@ -17,10 +17,22 @@ The types of changes are:
 
 ## [Unreleased](https://github.com/ethyca/fides/compare/2.9.0...main)
 
+### Added
+* Added Attentive erasure email connector [#2782](https://github.com/ethyca/fides/pull/2782)
+
+### Changed
+* Removed dataset based email connectors [#2782](https://github.com/ethyca/fides/pull/2782)
+* Changed Auth0's authentication strategy from `bearer` to `oauth2_client_credentials` [#2820](https://github.com/ethyca/fides/pull/2820)
+* renamed the privacy declarations field "Privacy declaration name (deprecated)" to "Processing Activity"
+
 
 ### Fixed
 * Fixed issue where the scopes list passed into FidesUserPermission could get mutated with the total_scopes call [#2883](https://github.com/ethyca/fides/pull/2883)
 * Restricted Contributors from being able to create Owners [#2888](https://github.com/ethyca/fides/pull/2888)
+
+### Removed
+
+* removed the `privacyDeclarationDeprecatedFields` flag
 
 ## [2.9.0](https://github.com/ethyca/fides/compare/2.8.3...2.9.0)